% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CDR_jobs.R
\name{calculate_cdr_jobs}
\alias{calculate_cdr_jobs}
\title{Calculate CDR Jobs}
\usage{
calculate_cdr_jobs(
  db_path,
  db_name,
  dat_file,
  scenario_list = NULL,
  region_list = NULL,
  output_path = "results",
  output_type = "csv",
  create_plots = TRUE,
  job_metric = "mean",
  ncol = 2,
  nrow = 2
)
}
\description{
Calculate CDR Jobs
}
