% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CDR_Job_Inten}
\alias{CDR_Job_Inten}
\title{Job Intensities for Various CDR Technologies}
\format{
A data frame with 4 columns:
\describe{
\item{main_technology}{Main CDR technology.}
\item{sub_technology}{Sub-technologies under each main technology.}
\item{min_int}{Minimum job intensity (jobs/tCO2).}
\item{mean_int}{Mean job intensity (jobs/tCO2).}
\item{max_int}{Maximum job intensity (jobs/tCO2).}
}
}
\usage{
CDR_Job_Inten
}
\description{
A dataset containing job intensities (\code{min_int}, \code{mean_int}, \code{max_int}) for various
Carbon Dioxide Removal (CDR) technologies.
}
\keyword{datasets}
